// frontend/app/src/app/core/api/model/monitorSummaryResponse.ts
/**
 * FS Bus API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface MonitorSummaryResponse {
    user_id: string;
    user_name?: string | null;
    user_surname?: string | null;
    full_name?: string | null;
    email?: string | null;
    role?: string | null;
    shift_count?: number;
    inspection_count?: number;
    failed_inspection_count?: number;
    last_shift_at?: string | null;
}