// frontend/app/src/app/core/api/model/selfieResponse.ts
/**
 * FS Bus API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Read model for one stored shift selfie.
 */
export interface SelfieResponse { 
    id: number;
    shift_id: number;
    timestamp: string;
    lat: number;
    lon: number;
    photo: string;
    created_at: string;
}